fix(fixture): Remove fixture from scene and venue
When deleting a fixture, it will now also be removed from the related scenes and venues.

fix #83
