+++ src/selectors/index.js
@@ -1,6 +1,8 @@
 import { createSelector } from 'reselect/src/index.js'
 import { collator } from '../utils/index.js'
 
+export * from './venue.js'
+
 /*
  * Selectors help you to retrieve data from the state so you don't have to write the
  * same code to access the state over and over again. It also helps to have a central position
@@ -30,7 +32,6 @@ export const getDekk = state => state.dekkManager
 export const getDekkConnected = state => state.dekkManager.connected
 export const getDekkData = state => state.dekkManager.data
 export const getUsbDmxControllerConnected = state => state.connectionManager.usb.connected
-export const getVenues = state => state.venueManager
 
 export const getAnimation = (state, properties) => {
   return getAnimations(state)
@@ -78,6 +79,14 @@ export const getScenesWithAnimation = (state, properties) => {
 }
 
 /*
+ * Get scenes that contain a certain fixtureId
+ */
+export const getScenesWithFixture = (state, properties) => {
+  return getScenes(state)
+    .filter(scene => scene.fixtures.includes(properties.fixtureId))
+}
+
+/*
  * Get a specific fixture by using the fixtureId
  */
 export const getFixture = (state, properties) => {
@@ -92,22 +101,6 @@ export const getFixtureByName = (state, properties) => {
   return getFixtures(state)
     .filter(fixture => fixture.name === properties.name)[0]
 }
-
-/*
- * Get a specific fixture by using the fixtureId
- */
-export const getVenue = (state, properties) => {
-  return getVenues(state)
-    .filter(venue => venue.id === properties.venueId)[0]
-}
-
-/*
- * Sort venues by venue.name
- */
-export const getVenuesSorted = createSelector(
-  getVenues,
-  venues => venues.sort((a, b) => collator.compare(a.name, b.name))
-)
 
 /*
  * Sort fixtures by fixture.name
